chore: deprecate markdown component (#1435)
