+++ docs/components/content/ExampleTheTitle.vue
@@ -1,5 +1,5 @@
 <template>
   <h1 class="text-4xl">
-    <Markdown :use="$slots.default" unwrap="p" />
+    <ContentSlot :use="$slots.default" unwrap="p" />
   </h1>
 </template>
