+++ docs/content/3.guide/1.writing/3.mdc.md
@@ -16,7 +16,7 @@ Any component in the `components/content/` directory or [made available globally
 The component must contain either:
 
 - A `<slot />`{lang="html"} to accept raw text or another component.
-- The `<Markdown />`{lang="html"} component to accept formatted text.
+- The `<ContentSlot />`{lang="html"} component to accept formatted text.
 
 In a markdown file, use the component with the **`::`** identifier.
 
@@ -107,7 +107,7 @@ You can add more `::::` when nesting components as a visual reminder.
 
 ### Markdown rendering
 
-The `<Markdown />`{lang="html"} component is auto-imported by Nuxt Content. It acts as a special slot that accepts rich text rendered by Markdown.
+The `<ContentSlot />`{lang="html"} component is auto-imported by Nuxt Content. It acts as a special slot that accepts rich text rendered by Markdown.
 
 The `unwrap` prop accepts an HTML tag that will be used to unwrap the content, useful when using tags such as title tags (`<h1>`{lang="html"}, `<h2>`{lang="html"}, ...) or inline tags (`<button>`{lang="html"}, `<a>`{lang="html"}, ...).
 
@@ -115,7 +115,7 @@ The `unwrap` prop accepts an HTML tag that will be used to unwrap the content, u
   ```html [TheTitle.vue]
   <template>
     <h1 class="text-4xl">
-      <Markdown :use="$slots.default" unwrap="p" />
+      <ContentSlot :use="$slots.default" unwrap="p" />
     </h1>
   </template>
   ```
@@ -131,15 +131,15 @@ The `unwrap` prop accepts an HTML tag that will be used to unwrap the content, u
   ::
 ::
 
-The `<Markdown />` component can act as a named slot with the `use` property:
+The `<ContentSlot />` component can act as a named slot with the `use` property:
 
 ```html
-<Markdown :use="$slots.description" unwrap="p">
+<ContentSlot :use="$slots.description" unwrap="p">
 ```
 
 ## Inline components
 
-Inline components are components without slots or `<Markdown />`.
+Inline components are components without slots or `<ContentSlot />`.
 
 They can be used with the `:` identifier.
 
@@ -189,7 +189,7 @@ defineProps(['type'])
 
 <template>
   <div :class="[type]">
-    <Markdown :use="$slots.default" unwrap="p" />
+    <ContentSlot :use="$slots.default" unwrap="p" />
   </div>
 </template>
   ```
