+++ docs/content/4.api/2.composables/3.unwrap.md
@@ -2,4 +2,4 @@
 title: 'useUnwrap()'
 ---
 
-It can be used to achieve a similar behavior as `unwrap`{lang=ts} prop from `<Markdown>`{lang=html} component.
+It can be used to achieve a similar behavior as `unwrap`{lang=ts} prop from `<ContentSlot>`{lang=html} component.
