+++ docs/content/5.examples/3.mdc/2.nested-components.md
@@ -4,7 +4,7 @@ title: 'Nested components'
 description: 'The MDC syntax allow you to nest components within a parent slot using indentation and the :: syntax.'
 ---
 
-The `components/AppNested.vue` component uses the `<Markdown>` component as markdown-rendering slot.
+The `components/AppNested.vue` component uses the `<ContentSlot>` component as markdown-rendering slot.
 
 ::ReadMore{link="/guide/writing/mdc#nesting"}
