+++ examples/mdc/nested-components/components/AppNested.vue
@@ -1,6 +1,6 @@
 <template>
   <aside>
     <h2>Nested component</h2>
-    <Markdown unwrap="p" />
+    <ContentSlot unwrap="p" />
   </aside>
 </template>
