+++ test/fixtures/basic/components/content/Alert.vue
@@ -1,5 +1,5 @@
 <template>
   <div>
-    <Markdown :use="$slots.default" />
+    <ContentSlot :use="$slots.default" />
   </div>
 </template>
