+++ test/fixtures/basic/components/content/TestMarkdown.vue
@@ -1,14 +1,14 @@
 <template>
   <div>
     <h1 class="text-3xl font-bold text-primary-900 dark:text-primary-100">
-      <Markdown :use="$slots.title" unwrap="p">
+      <ContentSlot :use="$slots.title" unwrap="p">
         Default title
-      </Markdown>
+      </ContentSlot>
     </h1>
     <p>
-      <Markdown :use="$slots.default" unwrap="p">
+      <ContentSlot :use="$slots.default" unwrap="p">
         Default paragraph
-      </Markdown>
+      </ContentSlot>
     </p>
   </div>
 </template>
